requirements.txt
# PIP
pip >= 20.1.1

# Telegram connection
telethon ~= 1.16.1, >= 1.16.1, < 1.16.4

# Config
pythondialog ~= 3.5.1

# Heroku & Updater
heroku3 ~= 4.2.1
gitpython ~= 3.1.9

# Translations
babel >= 2.8.0

# Eval
meval >= 2.4, ~= 2.4

# Web
aiohttp_jinja2 ~= 1.2.0

# # # Additional modules # # #

# Syntax highlighter
Pygments >= 2.7.4

# Video editor
moviepy >= 1.0.3

# Telegram stickers editor
lottie >= 0.6.6

# Python Imaging Library (Fork)
Pillow >= 8.1.0

# SVG converter
CairoSVG >= 2.5.1

# Youtube searcher
youtube-search >= 1.1.1

# Youtube downloader
youtube-dl2 >= 2.0

# Audio editor
pydub >= 0.24.1

# Image editor
Wand >= 0.6.5

# Search module
search-engine-parser >= 0.6.2

# GSBL
get-stick-bugged-lol >= 1.0.1

# QRcode
qrcode >= 6.1

